pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fieldflow"
version = "0.1.0"
description = "Generate FieldFlow servers that expose OpenAPI REST APIs as field-filtered tools for HTTP or MCP clients."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "Guillaume Gay", email = "opensource@guillaumegay.com" }
]
keywords = ["mcp", "fastapi", "openapi", "proxy", "llm"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
  "fastapi>=0.111.0",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.3.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
]

mcp = [
  "mcp>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["fieldflow*", "fieldflow_mcp*"]

[project.scripts]
fieldflow = "fieldflow.cli:main"
fieldflow-mcp = "fieldflow_mcp.cli:main"
